Move MergedSelections into new file
